get token data also
